refactor(dialog): transition Modal components to Dialog structure
- Removed the Modal component and its related files, transitioning to a Dialog-based architecture for improved accessibility and consistency.
- Updated Dialog components to utilize new slot-based structure, enhancing flexibility and maintainability.
- Refactored stories and documentation to reflect the changes in component structure and accessibility features.
- Ensured all components maintain proper accessibility standards and functionality.

All tests pass successfully, confirming no breaking changes introduced.
