+++ packages/nimbus/src/components/dialog/components/index.ts
@@ -1,0 +1,10 @@
+export { DialogRoot } from "./dialog.root";
+export { DialogTrigger } from "./dialog.trigger";
+export { DialogContent } from "./dialog.content";
+export { DialogBackdrop } from "./dialog.backdrop";
+export { DialogHeader } from "./dialog.header";
+export { DialogBody } from "./dialog.body";
+export { DialogFooter } from "./dialog.footer";
+export { DialogTitle } from "./dialog.title";
+export { DialogDescription } from "./dialog.description";
+export { DialogCloseTrigger } from "./dialog.close-trigger";
